pyproject.toml
[tool.poetry]
name = "notion-integration"
version = "0.1.0"
description = "Framework de integração e análise para Notion"
authors = ["The AI/OBoss <the@aioboss.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^24.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"